So much stuff
Created a WindowComm class to handle simple interactions with the user
that only require a yes or no, or simply a message for the user.

Designed the edit scene window and added functionality to a few of the
buttons (all buttons have action handlers, but not all action handlers
actually do something useful).

Clicking add scene or edit scene will open the scene window and disable
buttons on the SceneManagerPanel so that multiple are not opened.

Opening the edit window will load necessary information from the
selected scene into the window.

Added an ArrayList for connected scenes within the Scene class, also
added an ArrayList of labels that corresponds to the connected scenes.
(Each connected scenes label will be the message displayed to the player
of the game to move to that scene).

Have not yet added error checking to ensure that every scene has a label
attached.

Tyrone