New class, A lot of functionality added
Added a new class: AddConnectionWindow - used for when a scene is being
edited and the user has clicked on connect a scene.  Displays a list of
all scenes, the user may choose one.

Added a lot of functionality to EditSceneWindow.
Adding a scene will now actually add a scene, you can remove scenes and
edit selected scenes.

Current known bug:  Editing a scene and selecting save causes the
manager to think it is adding a new scene that has been edited into the
system.  This will cause a name conflict, as it is conflicting with
itself.  To fix, just make the scene manager check to see if the
requested scene to add is already contained within the list.  If it is,
check all names except for its own.

Added a few functions to scene, each scene has a flag now that
determines whether or not it has connections to it.  Later functionality
will check to see if it can actually be reached.

Various other bug fixes, probably various other bug creations too.

-Tyrone