Updated Scene/SceneManager/EditSceneWindow
EditSceneWindow:

Can now edit labels for connected scenes.
Can now remove connections from a scene

Scene:
Added a modifyTagByIndex function that takes in the index of the label
being modified and what it was modified to.  Changes the tag to the
specified label.

SceneManager:

Editing a scene and pressing save won't cause a name conflict

Tyrone