Lots of changes
Added the ability to add items to a scene.  Functionality not yet
available, but user will be able to choose an "unlock item" that the
player will require to enter the scene or a "drop item" that drops on
the scene, which will be added to the players inventory as soon as he
views the scene.

Added an additional column to the SceneTableModel that shows the names
of the items (if any) that is connected to the scene.  Updated the view
of the SceneTableModel to be more friendly with the data types received.

Various bug fixes.

Tyrone