Scene Window Changes/Additions
Item: Added a toString for easy JList viewing

Scene: Cleaned up getters and setters so they are more readable.

AddItemWindow: Added a drop down menu that allows the user to choose the
method in which the item will be added to the scene.  To do:
Implementation of the add item window.

EditSceneWindow: Added functionality for the Add Item button, which
opens the AddItemWindow.  Also added a function that AddItemWindow can
call to indicate to EditSceneWindow that it has closed, re-enabling the
Add Item Button.

MainWindow: Not sure, but it says I editted it.

- Tyrone