Updated SM, SMP, ESW, and AIW
SceneManager -
Now contains a start scene and end scene.  These scenes are not
removable from the system and their names are not allowed to be edited.
After each addition/removal/edit, the SceneManager now checks to see
what scenes are reachable from the start scene and updates the graph
accordingly.

SceneManagerPanel - If the user attempts to remove a start or end scene,
the user is informed that they cannot do that.

EditSceneWindow
-If the user edits the start or the end scene, the title window disables
editing.
-When the user chooses to save a scene, the title is now crosschecked
with the database before saving it to the scene.   (Before the name
would be saved within the scene, then cross-referenced with the
database).
-The Add Item button now opens a fully functional AddItemWindow.  The
X's next to each item will remove the item that has been added to the
window (if an item is connected to the slot).

AddItemWindow
- Allows the user to select and item and choose whether they want the
item to drop on the scene or unlock the scene based on a drop down menu
choice.
- An item may only be dropped on one Scene, if the user tries to add the
item to multiple scenes, they are informed with a meaningful error
message.
- An item may not be dropped on the scene that it unlocks.  If the user
attempts to do this, they are informed with a meaningful error message.

Except for minor aesthetic and design choices, SceneManager,
EditSceneWindow, Scene, AddItemWindow, and AddConnectionWindow should
all be ready for implementation.  Any changes within the editor at this
point should not affect the running of the game.

-Tyrone